Finish modularizing elections js page
