+++ fec/fec/static/js/modules/columns.js
@@ -675,66 +675,6 @@ var audit = [
   },
 ];
 
-function makeCommitteeColumn(opts, context, factory) {
-  return _.extend({}, {
-    orderSequence: ['desc', 'asc'],
-    className: 'column--number',
-    render: columnHelpers.buildTotalLink(['receipts', 'individual-contributions'], function(data, type, row, meta) {
-      row.cycle = context.election.cycle;
-      var column = meta.settings.aoColumns[meta.col].data;
-      return _.extend({
-        committee_id: (context.candidates[row.candidate_id] || {}).committee_ids,
-        two_year_transaction_period: row.cycle,
-      }, factory(data, type, row, meta, column));
-    })
-  }, opts);
-}
-
-var makeSizeColumn = _.partial(makeCommitteeColumn, _, _, function(data, type, row, meta, column) {
-  return columnHelpers.getSizeParams(column);
-});
-
-function sizeColumns(context) {
-  return [
-    {
-      data: 'candidate_name',
-      className: 'all',
-      width: 'column--med',
-      render: function(data, type, row, meta) {
-        return columnHelpers.buildEntityLink(
-          data,
-          helpers.buildAppUrl(['candidate', row.candidate_id]),
-          'candidate'
-        );
-      }
-    },
-    makeSizeColumn({data: '0'}, context),
-    makeSizeColumn({data: '200'}, context),
-    makeSizeColumn({data: '500'}, context),
-    makeSizeColumn({data: '1000'}, context),
-    makeSizeColumn({data: '2000'}, context)
-  ];
-}
-
-function stateColumns(results, context) {
-  var stateColumn = {'data': 'state'};
-  var columns = _.map(results, function(result) {
-    return makeCommitteeColumn(
-      {data: result.candidate_id},
-      context,
-      function(data, type, row, meta, column) {
-        return {
-          contributor_state: row.state,
-          committee_id: (context.candidates[column] || {}).committee_ids,
-          is_individual: 'true'
-        };
-      }
-    );
-  });
-
-  return [stateColumn].concat(columns);
-}
-
 module.exports = {
   candidateColumn: candidateColumn,
   committeeColumn: committeeColumn,
@@ -755,7 +695,5 @@ module.exports = {
   receipts: receipts,
   reports: reports,
   loans: loans,
-  audit: audit,
-  sizeColumns: sizeColumns,
-  stateColumns: stateColumns
+  audit: audit
 };
